Split backbone from image encoder
