+++ otx/algorithms/visual_prompting/adapters/pytorch_lightning/models/backbone/__init__.py
@@ -1,0 +1,8 @@
+"""Backbones for visual prompting model."""
+
+# Copyright (C) 2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .vit import ViT, build_vit
+
+__all__ = ["ViT", "build_vit"]
