+++ otx/algorithms/visual_prompting/adapters/pytorch_lightning/models/encoders/__init__.py
@@ -3,7 +3,7 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .sam_image_encoder import SAMImageEncoderViT
+from .sam_image_encoder import SAMImageEncoder
 from .sam_prompt_encoder import SAMPromptEncoder
 
-__all__ = ["SAMImageEncoderViT", "SAMPromptEncoder"]
+__all__ = ["SAMImageEncoder", "SAMPromptEncoder"]
