+++ otx/algorithms/visual_prompting/configs/sam_vit_b/config.yaml
@@ -15,7 +15,39 @@ dataset:
   
 model:
   name: SAM
-  backbone: vit_b
+  image_size: 1024
+  mask_threshold: 0.
+  return_logits: false
+  image_encoder:
+    backbone:
+      name: vit_b
+      patch_size: 16
+      depth: 12
+      embed_dim: 768
+      mlp_ratio: 4
+      num_heads: 12
+      qkv_bias: true
+      use_rel_pos: true
+      global_attn_indexes:
+      - 2
+      - 5
+      - 8
+      - 11
+      window_size: 14
+    freeze: true
+  prompt_encoder:
+    prompt_embed_dim: 256
+    mask_in_chans: 16
+    freeze: true
+  mask_decoder:
+    num_multimask_outputs: 3
+    transformer_cfg:
+      depth: 2
+      mlp_dim: 2048
+      num_heads: 8
+    iou_head_depth: 3
+    iou_head_hidden_dim: 256
+    freeze: false
   checkpoint: https://dl.fbaipublicfiles.com/segment_anything/sam_vit_b_01ec64.pth
 
 callback:
